Move TIMESTAMP TZ to TIMESTAMP cast tests to TestTimestampWithTimeZone class
They were in a separate class for when we had to parameterize the tests
for legacy vs new timestamp semantics. That's no longer necessary.
